requirements.txt
req
absl-py==0.11.0
aiobotocore
aiohttp==3.7.2
aioitertools==0.7.1
alembic==1.4.1
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs
backcall==0.2.0
base58==2.0.1
bleach
blinker==1.4
boto3
botocore
branca==0.4.1
bs4==0.0.1
cachetools==4.1.1
category-encoders==2.2.2
certifi==2020.6.20
cffi==1.14.3
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
cmdstanpy==0.9.5
confuse==1.3.0
convertdate==2.2.2
coverage==5.3
cryptography==3.2.1
cycler==0.10.0
Cython==0.29.17
databricks-cli==0.14.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
ephem==3.7.7.1
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.4
folium==0.11.0
fsspec==0.8.4
future==0.18.2
gast==0.3.3
gcsfs==0.6.0
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.11
gorilla==0.3.0
grpcio==1.33.2
gunicorn==20.0.4
h5py==2.10.0
herepy==2.1.1
holidays==0.10.3
htmlmin==0.1.12
identify==1.5.9
idna==2.10
ImageHash==4.1.0
imageio==2.9.0
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata==2.0.0
iniconfig==1.0.1
ipykernel
ipython
ipython-genutils
ipywidgets
isodate==0.6.0
isort==5.5.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
json5==0.9.5
jsonschema==3.2.0
jupyter
jupyter-client
jupyter-console
jupyter-core
jupyterlab-pygments
kiwisolver==1.2.0
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.4.3
llvmlite==0.34.0
LunarCalendar==0.0.9
lxml==4.5.2
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
memoized-property==1.0.3
missingno==0.4.2
mistune==0.8.4
more-itertools
msrest==0.6.19
multidict==5.0.2
nbconvert
nbformat
nest-asyncio==1.4.1
networkx==2.5
nodeenv==1.5.0
notebook
numba==0.51.2
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
patsy==0.5.1
pexpect==4.8.0
phik==0.10.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
pmdarima==1.7.1
pre-commit==2.8.2
prometheus-client==0.8.0
prometheus-flask-exporter==0.18.1
prompt-toolkit==3.0.7
protobuf==3.13.0
psutil==5.7.0
ptyprocess==0.6.0
py==1.9.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.5.0
Pygments==2.7.1
pylint==2.6.0
PyMeeus==0.3.7
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole
querystring-parser==1.2.4
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
s3fs==0.5.1
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn
Send2Trash==1.5.0
setuptools-git==1.2
simplejson==3.17.2
six==1.15.0
smart-open==3.0.0
smmap==3.0.4
soupsieve==2.0.1
spotipy==2.16.1
tabulate==0.8.7
tangled-up-in-unicode==0.0.6
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
tox==3.20.1
traitlets
typing-extensions==3.7.4.3
urllib3==1.25.10
validators==0.18.1
virtualenv==20.1.0
visions==0.5.0
watchdog==0.10.3
wcwidth
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
XlsxWriter==1.3.7
zipp==3.3.0
